chore(*): remove typing_trigger
This feature was essentially a workaround for the disable-while-typing feature not functioning properly in libinput,
rather than a keypress function. As a result, this code can be removed.

The use of a typing trigger for a simulated disable-while-typing is no longer necessary.

In the case of external keyboards with libinput,
the disable-while-typing feature does not work.
Key remappers are often treated as external keyboards,
which can cause accidental taps.

Following code uses libinput quirks to treat xremap's virtual keyboard as an internal keyboard.

/etc/libinput/local-overrides.quirks
```
[xremap for disable-touch-while-typing]
MatchUdevType=keyboard
MatchName=xremap*
AttrKeyboardIntegration=internal
```